fix: move test case from button to ensembleAction
